+++ core/src/Streamly/Internal/Data/Serialize.hs
@@ -12,7 +12,19 @@
 --
 
 module Streamly.Internal.Data.Serialize
-    ( module Streamly.Internal.Data.Serialize.Type
+    (
+    -- * Unbox
+      module Streamly.Internal.Data.Unbox
+    , module Streamly.Internal.Data.Unbox.TH
+    -- * Unboxed IORef
+    , module Streamly.Internal.Data.IORef.Unboxed
+    -- * Serialize
+    , module Streamly.Internal.Data.Serialize.Type
+    -- * Serialize TH
+    , module Streamly.Internal.Data.Serialize.TH
+    , module Streamly.Internal.Data.Serialize.TH.RecHeader
+    , module Streamly.Internal.Data.Serialize.TH.Common
+    , module Streamly.Internal.Data.Serialize.TH.Bottom
     ) where
 
 --------------------------------------------------------------------------------
@@ -21,7 +33,6 @@ module Streamly.Internal.Data.Serialize
 
 import Data.Proxy (Proxy(..))
 import Streamly.Internal.Data.Array (Array(..))
-import Streamly.Internal.Data.Unbox (MutableByteArray(..))
 import GHC.Exts (Int(..), sizeofByteArray#, unsafeCoerce#)
 import GHC.Word (Word8)
 
@@ -32,6 +43,13 @@ import GHC.Integer.GMP.Internals (Integer(..), BigNat(..))
 #endif
 
 import Streamly.Internal.Data.Serialize.Type
+import Streamly.Internal.Data.Unbox
+import Streamly.Internal.Data.Unbox.TH
+import Streamly.Internal.Data.Serialize.TH
+import Streamly.Internal.Data.Serialize.TH.RecHeader
+import Streamly.Internal.Data.Serialize.TH.Common
+import Streamly.Internal.Data.Serialize.TH.Bottom
+import Streamly.Internal.Data.IORef.Unboxed
 
 --------------------------------------------------------------------------------
 -- Common instances